Make VGG code compatible with old/new format data
Commit 6197b175bf1f15d757b1d92f6747c5faa06c1ed8 noted that the VGG model
data was available in multiple formats. The code was only compatible
with the old format, so the README was updated to include a link to the
old format data. This was a source of confusion, because people didn't
download the model using the link from the README and ended up with the
new-format data, which caused the program to crash. Later, a more
descriptive error message was added, but this continued to be a source
of confusion.

This patch updates the VGG data loading code to be compatible with both
old and new data formats.