Fix format detection with Python 2
Format detection didn't work correctly with Python 2 because with the
new format data, `weights[i][0][0][0][0]` was an instance of
`numpy.unicode_`, which is not a subclass of `str`. The check worked
fine with Python 3, where the value was an instance of `numpy.str_`,
which is a subclass of `str`.

This patch changes the way we detect the data format so it works on both
Python 2 and Python 3.